Did a lot of work with economy and shops and minebay. Made it to where amounts and fields cannots surpass Integer.MAX_VALUE. Also made balance for ATM long and getting cash pile balances LONG.

+++ main/java/com/minelife/economy/client/gui/atm/GuiATMBase.java
@@ -19,9 +19,9 @@ public class GuiATMBase extends GuiScreen {
     protected int block = 0;
 
     public int guiLeft, guiTop, xSize = 176, ySize = 186;
-    public int balance;
+    public long balance;
 
-    public GuiATMBase(int balance) {
+    public GuiATMBase(long balance) {
         this.balance = balance;
     }
